Refactor font rendering out of main.cpp
